fix paths issues raised after moving some components to a new foder
